Image modifications.
CATSFC/system/gui/boot.bmp: Remove the Engrish, although it was funny ("Enjoy yourself! All pleasure in it"). Replace it with a controller and the text "Loading...".

catsfc.bmp: Remove the CAT and replace it with a controller. In the .ini, it's already written as CATSFC anyway.

copyright: Update with image attribution per CC-BY-NC-ND 3.0.

source/images: Create, with .psd files for the two modified .bmp files.